==> script_pelicula/01-generar_usuarios.ipynb <==
# Generar Usuarios

import pandas as pd
from faker import Faker
import random

fake = Faker('es_ES')  # o 'es_ES' 

rango = range(1,138494)

data = []
userid = 0
for i in rango:
    nombre = fake.name()
    email = nombre.lower().replace(" ", ".") + "@ejemplo.com"
    # edad = random.randint(18, 80)
    genero = random.choice(['M', 'F'])
    # nogenre = random.choice([1, 0])
    # action = random.choice([1, 0])
    # adventure  = random.choice([1, 0])
    # animation = random.choice([1, 0])
    # children = random.choice([1, 0])
    # comedy = random.choice([1, 0])
    # crime = random.choice([1, 0])
    # documentary = random.choice([1, 0])
    # drama = random.choice([1, 0])
    # fantasy = random.choice([1, 0])
    # film_noir = random.choice([1, 0])
    # horror = random.choice([1, 0])
    # imax = random.choice([1, 0])
    # musical   = random.choice([1, 0])
    # mystery = random.choice([1, 0])
    # romance = random.choice([1, 0])
    # sci_fi = random.choice([1, 0])
    # thriller = random.choice([1, 0])
    # war_western  = random.choice([1, 0])

    provincia = random.choice( [
        "Azuay",
        "Bolívar",
        "Cañar",
        "Carchi",
        "Chimborazo",
        "Cotopaxi",
        "El Oro",
        "Esmeraldas",
        "Galápagos",
        "Guayas",
        "Imbabura",
        "Loja",
        "Los Ríos",
        "Manabí",
        "Morona Santiago",
        "Napo",
        "Orellana",
        "Pastaza",
        "Pichincha",
        "Santa Elena",
        "Santo Domingo de los Tsáchilas",
        "Sucumbíos",
        "Tungurahua",
        "Zamora Chinchipe"
    ])
    
    # fila = [i, nombre, email, edad, genero, provincia, nogenre,action,adventure,animation,children,comedy,crime,documentary,drama,fantasy,film_noir,horror,imax,musical,mystery,romance,sci_fi,thriller,war_western]
    fila = [i, nombre, email, genero, provincia]
    data.append(fila)

# columnas = ['userid', 'nombre', 'email', 'edad', 'genero', 'provincia', 'nogenre','action','adventure','animation','children','comedy','crime','documentary','drama','fantasy','film_noir','horror','imax','musical','mystery','romance','sci_fi','thriller','war_western'] 
columnas = ['userid', 'nombre', 'email', 'genero', 'provincia'] 

df = pd.DataFrame(data, columns=columnas)

df.head()

def sin_tilde(email):
    replacements = (("á", "a"), ("é", "e"), ("í", "i"), ("ó", "o"), ("ú", "u"))
    for a, b in replacements:
        email = email.replace(a, b).replace(a.upper(), b.upper())
    return email

df['email'] = df['email'].apply(sin_tilde)

df.head()

# Guardado de tabla dimension usuarios temporal
df.to_csv(r"../data/usuarios.csv")
