==> script_pelicula/02-rating-Version2.ipynb <==
# Rating

import pandas as pd
import missingno as msno
import numpy as np

rating = pd.read_csv(r"../data/rating.csv")

rating.columns = rating.columns.str.lower() # en minuscular
rating.columns = rating.columns.str.strip() # quitar espacios en blanco
rating.head()

# Total de registros
len(rating)

# Verificamos que no haya nulos
rating.isna().sum()

# Cambiar el tipo de datos `object` de timestamp a datetime
rating["timestamp"] = pd.to_datetime(rating["timestamp"])

rating.head()

rating.info()

# Adicionamos la columna `year` y llenar con el año del timestamp
rating["year"] =  rating['timestamp'].dt.year
rating["month"] =  rating['timestamp'].dt.month
rating.head()

# DATA OPCION 1

Promedio y conteo en forma general agrupado unicamente por movieid

# Obtenemos el rating promedio y el conteo general 
rating_movies_promedio = rating.groupby(["movieid"])["rating"].mean()
rating_movies_conteo = rating.groupby(["movieid"])["rating"].count()


# Promedio de rating por pelicula de forma global en todos los años
rating_movies_promedio.head()

# Conteo de rating por pelicula de forma global en todos los años
rating_movies_conteo.head()

# union de promeido y conteo
rating_promedio_conteo = pd.merge(rating_movies_promedio, rating_movies_conteo, on="movieid", how="left")

rating_promedio_conteo.head()

# renombrar columnas
rating_promedio_conteo = rating_promedio_conteo.rename(
    columns={
        "rating_x": "rating_promedio",
        "rating_y": "rating_conteo"
    }
)

rating_promedio_conteo.head()

# Es el total de registros
rating_promedio_conteo["rating_conteo"].sum()
# Esta OK

rating_promedio_conteo.head()

# Guardado de tabla dimension usuarios temporal
rating_promedio_conteo.to_csv(r"../data/partial_rating.csv")

# DATA OPCION 2
Promedio agrupado por movieid y year
Se procede a crear una tabla que las columnas sean los años y ahi el promedio 
Resultado -> archivo tipo insumo para tabla de hecho `partial_rating_year.csv`

df_year_promedio = rating.groupby(["movieid","year"]).agg({"rating": "mean"})

# df_year_promedio.reset_index(inplace=True)
# df_year_promedio.set_index("movieid", inplace=True)
df_year_promedio

type(df_year_promedio)

df_pivot = df_year_promedio.pivot(columns='year', values=['rating'])

df_pivot.head()

df_pivot = df_pivot.replace(np.nan, 0)

df_pivot.head()

type(df_pivot)

df_pivot = df_pivot.reset_index()

df_pivot.head()

df_pivot.rename(columns={'movieid': 'movieid'}, inplace=True) 

df_pivot.head()

columns = [col[1] if isinstance(col, tuple) else col for col in df_pivot.columns]
columns[0] = "movieid"
df_pivot.columns = columns

year_cols = [c for c in df_pivot.columns if isinstance(c, str) and c.isdigit()]
for col in year_cols:
    df_pivot[col] = pd.to_numeric(df_pivot[col], errors='coerce')

df_pivot.head()

df_pivot.reset_index()

df_pivot.set_index("movieid", drop=True)

# Guardado de tabla de dimension

df_pivot.to_csv(r"../data/partial_rating_year.csv")

# DATA OPCION 3
Promedio y conteo agrupado por movieid, userid, year, month
Resultado -> archivo tipo tabla de dimension

### Esto es para una tabla de dimensión `d_rating`
Contendra la pelicula agrupado por usuario, año y por mes

# Obtenemos el rating promedio por pelicula, año y mes
grupo = rating.groupby(["movieid","userid", "year","month"])["rating"]
rating_movies_year_month_promedio = grupo.mean()
rating_movies_year_month_conteo = grupo.count()

# Deberia ser una table dimension
rating_movies_year_month_promedio.head()

rating_movies_year_month_conteo.head()

# union de promeido y conteo
dim_rating = pd.merge(rating_movies_year_month_promedio, rating_movies_year_month_conteo, on=["movieid","userid", "year","month"], how="left")

dim_rating.head()

# renombrar columnas
dim_rating = dim_rating.rename(
    columns={
        "rating_x": "promedio",
        "rating_y": "conteo"
    }
)

dim_rating.reset_index(inplace=True)

dim_rating

dim_rating.head()

dim_rating["ratingid"] = range(1, len(dim_rating) + 1)

dim_rating.set_index("ratingid", inplace=True)

dim_rating.head()

# Guardado de tabla de dimension
dim_rating.to_csv(r"../data/d_rating.csv")
