==> script_pelicula/03-tags.ipynb <==
# Tags

import pandas as pd
import missingno as msno
import numpy as np

df = pd.read_csv(r"../data/tag.csv")

df.columns = df.columns.str.lower() # en minuscular
df.columns = df.columns.str.strip() # quitar espacios en blanco
df.head()

# Total de registros
len(df)

# Verificamos que no haya nulos
df.isna().sum()

# Cambiar el tipo de datos `object` de timestamp a datetime
df["timestamp"] = pd.to_datetime(df["timestamp"])

df.info()

# consultar columnas de valores ausentes en columnas dadas
df[df.tag.isna()]

# Eliminamos los nulos
df = df.dropna(subset=["tag"], how="all")

# Verificamos que no haya nulos
df.isna().sum()

len(df)

465564 - 465548

es_numero = pd.to_numeric(df['tag'], errors='coerce').notna()
df = df[~es_numero]
df

df_strings = df[df['tag'].apply(lambda x: isinstance(x, str))]

df_strings['tag'] = df_strings['tag'].str.replace(r'[^a-zA-Z0-9]', '', regex=True)

df_strings

def join_unique_tags(tags):
    return ', '.join(pd.unique(tags))

df = df_strings.groupby(['movieid'])['tag'].apply(join_unique_tags)

df.head()

# # Guardado de tabla de dimension
# df.to_csv(r"../data/partial_tags.csv")
