==> script_pelicula/04-movies.ipynb <==
# Movies

import pandas as pd
import missingno as msno
import numpy as np

df = pd.read_csv(r"../data/process/movie.csv")


df.columns = df.columns.str.lower() # en minuscular
df.columns = df.columns.str.strip() # quitar espacios en blanco
df.head()

# Total de registros
len(df)

# Verificamos que no haya nulos
df.isna().sum()

movie = df.copy()
genres_dummies = movie['genres'].str.get_dummies('|') # Separa generos por '|' y crear One-Hot Encoding

genres_dummies.head()

rating = pd.read_csv(r"../data/partial_rating.csv")

rating.head()

df = pd.merge(df, rating, on="movieid", how="left")

df.head()

tags = pd.read_csv(r"../data/partial_tags.csv")

df = pd.merge(df, tags, on="movieid", how="left")

df.head()

df = df.set_index('movieid')


df.head()

# Guardado de tabla de hecho
df.to_csv(r"../data/h_movie.csv")
