==> script_pelicula/09-pruebas_api.ipynb <==
import pandas as pd
import requests
import os
import numpy as np

def load_data(file_path):
    BASE_DIR = os.getcwd() 
    DATA_PROCESS_DIR = os.path.join(BASE_DIR, "..", 'data', 'process')
    PROCESSED_FILE = os.path.join(DATA_PROCESS_DIR, file_path)
    # print(PROCESSED_FILE)
    try:
        df = pd.read_csv(PROCESSED_FILE, encoding='utf8')
        return df
    except FileNotFoundError:
        print(f"Error: No se encontró el archivo {PROCESSED_FILE}")
        return None

orderby = "Rating"
genres = []
rating_min = 0.0
rating_max = 5.0
items_per_page = 10

data = load_data("movie_perfil_contenido.csv")

col_categoricas = ['movieid', 'title', 'genres', 'rating_promedio', 'rating_conteo', 'tag', 'tmdbid']
df_procesado = pd.DataFrame(data, columns=col_categoricas)
df_procesado.reset_index(level=0, inplace=True)

ddata = df_procesado.copy()

if len(genres) > 0:
    # ddata[ddata["genres"].str.split("|").apply(lambda x: )]}
    filtro = ddata['genres'].str.split("|").apply(lambda x: any(np.isin(genres, x)))
    ddata = ddata[filtro]

filtro = (ddata["rating_promedio"] >= rating_min) & (ddata["rating_promedio"] <= rating_max)
ddata = ddata[filtro]

# orderby = "Popularidad"
if orderby == "Rating":
    ddata = ddata.sort_values(["rating_promedio"], ascending=False)
if orderby == "Popularidad":
    ddata = ddata.sort_values(["rating_conteo"], ascending=False)
ddata = ddata.iloc[0:items_per_page]
ddata
