==> script_pelicula/10-MachineSVD.ipynb <==
# Singular Value Decomposition (SVD)

import pandas as pd
import numpy as np
from numpy.linalg import svd

# Cargar datos limpios
ratings_original = pd.read_csv(r"../data/process/clean_rating.csv")

ratings_original

ratings_original.shape

movies = pd.read_csv(r"../data/process/movie_perfil_contenido.csv",
        # usecols = ["movieid", "title"]
)
movies.shape

df = pd.merge(ratings_original, movies, on='movieid')

df.head()

n_users = df['userid'].nunique()
n_movies = df['movieid'].nunique()

print(f"Numero de users: {n_users}")
print(f"Numero de Peliculas: {n_movies}")
print(f"Total ratings: {len(df)}")

user_item_matrix = df.pivot_table(index='userid', columns='movieid', values='rating').fillna(0)

print("Matrix:", user_item_matrix.shape)

R = user_item_matrix.to_numpy()

U, sigma, Vt = svd(R, full_matrices=False)
print(f"Shapes -> U: {U.shape}, Sigma: {sigma.shape}, Vt: {Vt.shape}")

k = 50
sigma_k = np.diag(sigma[:k])

R_approx = np.dot(np.dot(U[:, :k], sigma_k), Vt[:k, :])
R_pred_df = pd.DataFrame(R_approx, 
                        index=user_item_matrix.index, 
                        columns=user_item_matrix.columns)

R_pred_df.head()

user_id = 31

# Original ratings and predictions
user_ratings = user_item_matrix.loc[user_id]
user_predictions = R_pred_df.loc[user_id]

# Recommend top 10 unrated movies
unrated = user_ratings[user_ratings == 0]
recommendations = user_predictions[unrated.index].sort_values(ascending=False).head(10)
recommendations.reset_index( inplace=False )
#recommendations

# rdf = pd.merge(recommendations, movies, left_on="movieid", right_on="movieid", how="left")
rdf = pd.merge(recommendations, movies, on="movieid", how="left")
rdf.head()
